refactor: perf calcs in own modules
